swap idle configs

+++ scripts/config_carousel/check-worker-availability.sh
@@ -21,9 +21,23 @@ check_availability() {
 }
 
 if [[ "$AVAILABLE" == "true" ]]; then
-    echo "Workers are available"
-    sudo systemctl start $QUIL_SERVICE_NAME
+    echo "Workers are available, switching to in-use config"
+    IN_USE_CONFIG_FILE=$(yq eval '.scheduled_tasks.config_carousel.in_use_config_file // "~/ceremonyclient/node/in-use-config.yml"' $QTOOLS_CONFIG_FILE)
+    if [ -f $IN_USE_CONFIG_FILE ]; then
+        cp $IN_USE_CONFIG_FILE $QUIL_NODE_CONFIG_FILE/.config/config.yml
+        sudo systemctl restart $QUIL_SERVICE_NAME
+    else
+        echo "No in-use config file found ($IN_USE_CONFIG_FILE), starting whatever is already defined"
+        sudo systemctl start $QUIL_SERVICE_NAME
+    fi
 else
-    echo "Workers are not available"
-    sudo systemctl stop $QUIL_SERVICE_NAME
+    echo "Workers are not available, switching to idle config"
+    IDLE_CONFIG_FILE=$(yq eval '.scheduled_tasks.config_carousel.idle_config_file // "~/ceremonyclient/node/idle-config.yml"' $QTOOLS_CONFIG_FILE)
+    if [ -f $IDLE_CONFIG_FILE ]; then
+        cp $IDLE_CONFIG_FILE $QUIL_NODE_CONFIG_FILE/.config/config.yml
+        sudo systemctl restart $QUIL_SERVICE_NAME
+    else
+        echo "No idle config file found ($IDLE_CONFIG_FILE), stopping node"
+        sudo systemctl stop $QUIL_SERVICE_NAME
+    fi
 fi
